🧹 [code health] Extract file search logic to shared helper
